+++ application/src/components/method-details/availability/details.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { useApplicationContext } from '@commercetools-frontend/application-shell-connectors';
 import messages from '../messages';
 import { useIntl } from 'react-intl';
@@ -178,12 +179,12 @@ const AvailabilityDetails = (props: TAvailabilityDetailFormProps) => {
   };
 
   const handleSubmit = async (formikValues: TAmountPerCountry) => {
-    let updateObject = Object.assign({}, props.method);
+    const updateObject = Object.assign({}, props.method);
 
     const newPricingConstraints = generatePricingConstraints(formikValues);
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    let clonedValue: any = Object.assign({}, updateObject.value);
+    const clonedValue: any = Object.assign({}, updateObject.value);
     clonedValue.pricingConstraints = newPricingConstraints;
 
     if (props.method?.container && props.method?.key && formikValues) {